+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 ### Added
 * `VectorSearchIndex` Databricks resource
 * `VectorSearchEndpoint` Databricks resource
+* `purge` method for data sink
+* `refresh` option for pipeline and pipeline node
 ### Fixed
 * Checkpoint location of `TableDataSink`
 ### Updated